+++ quote/YusufTansu.java
@@ -1,0 +1,18 @@
+package quote;
+
+public class YusufTansu implements Quotation {
+    public int year() {
+        return 1936;
+    }
+    public String text() {
+        String p = "I don't care that they stole my idea . . \n";
+        String s = "I care that they don't have any of their own. ";
+        return p+s;
+    }
+    public String author() {
+        return "Nikola Tesla";
+    }
+    public String reference() {
+        return "https://www.goodreads.com/author/quotes/278.Nikola_Tesla";
+    }
+}
